Deduce Vertex Collections for Traversal from Named Graphs (#21988)
* Deduplicate code

* Try without throwing...

* Cleanup

* Fix it

* GraphManager support to find VertexCollections

Introduces a method to the graph manager that allows finding which
vertex collections are attached to a given edge collection by
inspecting all graphs in the database to find whether any of them has
the given edge collection in it.

This code assumes that these definitions are consistent among graphs
if an edge collection is shared between multiple named graphs.

No tests yet.

* Add collections

* Move adding of collections around

* Add red test

* Add adding of vertex collections from named graphs

* Rename functions

* Try handling more cases

* Fix crash in IResearch tests

* Remove code duplication

* Add more tests

* eslint happiness

* Fix test for non-cluster

* Remove failing tests

* Undo gratuitous changes

* Undo formatting changes

* Cleanup

* Add CHANGELOG entry

* Fix after rebase

* Use some functions

* Handle another type of node

* Rewrite vertexCollection deduction code

* Cleanup

* Step

* Only add vertex collections when option not given

Only add vertrex collections for edge collections if the graph operation
uses edgeCollection syntax without giving a vertexCollections option

* Refactor node matching a bit

* More tests

* More tests and a fix

* Fix IResearch test