Remove old no-eslint files (#22102)
* Remove old no-eslint files

These files were originally added as copy of the package.json just without eslint. But when the packages were updated to newer versions the no-eslint files were not adjusted accordingly. However, in our build script we were replacing the package.json and package-lock.json with the no-eslint versions, so we ended up with those files referring to older versions while not packages in the node_modules folder actually contained newer versions.

Since in the meantime we have already removed eslint from the list anyway, we not longer need this special stunt. Also, this PR moves the node-netstat module to the devDependencies since we only use it in our internal tests and don't include it in release packages.

* Update CHANGELOG

---------

Co-authored-by: Vadim Kondratev <vadim@arangodb.com>